requirements.txt
asgiref==3.4.1
boto3==1.18.13
botocore==1.21.13
crispy-bootstrap5==0.4
dj-database-url==0.5.0
Django==3.2.5
django-crispy-forms==1.12.0
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
django-storages==1.11.1
docxtpl==0.11.5
gunicorn==20.1.0
Jinja2==3.0.1
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==2.0.1
phonenumbers==8.12.28
Pillow==8.3.1
psycopg2==2.9.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.0
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
